django/requirements.txt
Django==1.11.18
transifex-python==0.3.2
